WIP: update device service to new code generation
