+++ services/device/src/index.ts
@@ -5,9 +5,10 @@ import WebSocket from 'ws'
 import { config } from './config'
 import { AppDataSource } from './data_source'
 import { app } from './generated/index'
-import { isUserType, JWTVerificationError } from './generated/types'
+import { isUserType } from './generated/types'
 import { peerconnectionsCallbackHandling } from './methods/callbacks'
-import { deviceHandling } from './operations/devices'
+import { deviceHandling } from './operations'
+import { JWTVerificationError } from './types/errors'
 
 declare global {
     namespace Express {
@@ -17,36 +18,6 @@ declare global {
             wsListeners: Map<string, (socket: WebSocket) => void>
         }
     }
-}
-
-/**
- * This function attempts to verify the provided JWT.
- * @param jwt The JWT to be verified.
- * @param scopes The scopes needed to access the requested resource.
- * @throws {JWTVerificationError} Thrown
- * @returns The user to which the JWT belongs.
- */
-async function JWTVerify(jwt: string | undefined, scopes: string[]) {
-    if (!jwt) throw new JWTVerificationError('No JWT provided', 401)
-    if (!config.SECURITY_ISSUER)
-        throw new JWTVerificationError('No security issuer specified', 500)
-    const jwksUri = new URL(
-        config.BASE_URL.endsWith('/')
-            ? config.BASE_URL + '.well-known/jwks.json'
-            : config.BASE_URL + '/.well-known/jwks.json'
-    )
-    const JWKS = createRemoteJWKSet(jwksUri)
-    const jwtVerifyResult = await jwtVerify(jwt, JWKS, {
-        issuer: config.SECURITY_ISSUER,
-        audience: config.SECURITY_AUDIENCE,
-    })
-    if (!isUserType(jwtVerifyResult.payload))
-        throw new JWTVerificationError('Payload is malformed', 401)
-    const user = jwtVerifyResult.payload
-    for (const scope of scopes) {
-        if (user.scopes.includes(scope)) return user
-    }
-    throw new JWTVerificationError('Missing Scope: one of ' + scopes, 403)
 }
 
 AppDataSource.initialize()
@@ -66,7 +37,41 @@ AppDataSource.initialize()
             next()
         })
         app.initService({
-            JWTVerify: JWTVerify,
+            security: {
+                JWT: async (req, scopes) => {
+                    const authorization_header = req.header("Authorization")
+                    if (authorization_header === undefined) {
+                        throw new JWTVerificationError("Authorization header is not set", 401)
+                    }
+                    const bearerTokenResult = /^Bearer (.*)$/.exec(authorization_header);
+                    if (bearerTokenResult === null || bearerTokenResult.length != 2) {
+                        throw new JWTVerificationError("Authorization header is malformed", 401)
+                    }
+                    const jwt = bearerTokenResult[1]
+                    if (!jwt) throw new JWTVerificationError('No JWT provided', 401)
+                    if (!config.SECURITY_ISSUER)
+                        throw new JWTVerificationError('No security issuer specified', 500)
+                    const jwksUri = new URL(
+                        config.BASE_URL.endsWith('/')
+                            ? config.BASE_URL + '.well-known/jwks.json'
+                            : config.BASE_URL + '/.well-known/jwks.json'
+                    )
+                    const JWKS = createRemoteJWKSet(jwksUri)
+                    const jwtVerifyResult = await jwtVerify(jwt, JWKS, {
+                        issuer: config.SECURITY_ISSUER,
+                        audience: config.SECURITY_AUDIENCE,
+                    })
+                    if (!isUserType(jwtVerifyResult.payload))
+                        throw new JWTVerificationError('Payload is malformed', 401)
+                    const user = jwtVerifyResult.payload
+                    for (const scope of scopes) {
+                        if (user.scopes.includes(scope)) {
+                            return user
+                        }
+                    }
+                    throw new JWTVerificationError('Missing Scope: one of ' + scopes, 403)
+                }
+            }
         })
         peerconnectionsCallbackHandling(app)
         const wsServer = new WebSocket.Server({ noServer: true })
